Update the code to properly push out the generated graph files.
Make the new code the default.

It's been tested under FreeBSD and Linux.  Pls report problems to ntop-dev.

If you need to disable it, there is a switch in ntop.h, just uncomment:

/* #define DISABLE_GDC_WATCHDOG */


If the libpng conflict exists on your system, instead of crashing the ntop
web server, this code detects it and will push out an "error" graphic.

-----Burton (ref 149, part2)