Minor fixups beyond BMS0188...
Add explict test for autoconf version.
Move the automake/autoconf tests to the very top.
Test for awk up front.
Fixup autotoolversions test so it runs under awk as well as gawk.

Understood that this removes the (minute) possibility of having ./configure
work without the auto* tools installed at all, and/or a down level version.
That only worked through blind dumb luck before, this locks it down.

S'be'it...

-----Burton (ref 194)