When you process netFlow data, there's one major issue, which is that ntop doesn't see the contents of the packets, just the header.
Accordingly, ntop can't classify ftp traffic (where the
ftp server and client agree to use a random pair of ports)
since it doesn't see the PORT command and so there's nothing
for the netFlow plugin to classify.

This patch adds a ./configure parameter:

     --enable-netflowassumeftpdata

Which tells ntop to classify the unclassified data (that is
not netflow and ports > 1023) as ftp-data.

If you really, really understand your data, this might be
a good assumption.  But for most situations, this is
probably not a good choice.

Regardless, this patch adds some stats to the netFlow
plugin which will show you what packets were thrown
away.

-----Burton (ref 257)