BMS0407
 1. Remove the xmldump (non-working) functionality from the main ntop code.
    It's just too fragile both for development and for operation.

 2. Add a plugin, xmldumpPlugin.c which implements the ability to dump ntop
    data as an xml file.

The only pieces in the main code are the bare minimums...

   A few constants in globals-defines.h, URLsecurity for the unique (.xml and
   .dtd) files in the plugin and a MAKE_WITH_XMLDUMP constant in ntop.h.

Note that the setting of the (strange) places for the required libraries is now
in the configureextra files, and only LINUXredhat9 is actually populated.

If you don't have the libraries a minimal plugin is built, which tells you it
won't run.


This implies a change - instead of having the 'javadoc' type documentation
in globals-structtypes.h, where it's in your face to update it (but much,
much, harder to read), there's now a separate file, plugins/globals-structtypes.xml.


By doing this change, I can continue work on this piece of functionality without
having or causing issues to delay ntop 3.0

-----Burton