BMS0423
Update for html.

All of these have been through w3c validator and will validate
if you change the indicated meta tag from a comment.  For those
which simply won't validate, there are w3c compliant versions -
rename the xxx_w3c.html to xxx.html.

Note that there are a number of files that will be deleted as
they are no longer used.  I cleaned up and commited them so
the final version in the cvs would be 'good'.

This should <ha/>, <ha/> be the final cycle on these menus.
Please look 'em over.

-----Burton