Put back the --set-pcap-nonblocking option, which still appears to be required for SOME users of FreeBSD 4.x.
Unlike 3.0, we are no longer trying to be clever.  If you need it you
will have to specify it on the command line or via the web server
configuration.

According to the man page for pcap, this is not supposed to work,
but it does.  So it's unsupported...

Symptom:  When first attaching to the web server it hangs or takes
a very long time (10+ minutes) to view the first page.

Note: This will increase the cpu usage of ntop - read docs/FAQ.

-----Burton (ref 531)