Update p2c - this generates the IP to Country Code mapping file from the registry data.
(1) AfriNIC - new Registry for Africa.  Note that for April 2005, we force
    use of the 2005-04-15 (official launch date) file.

(2) Do not fail if the Cable & Wireless data file is missing.

Note that the C&W data appears to have disappeared.  This is bad.

Using current data (witout C&W) will cause a massive shift in IP locations
from various European countries to "UK", because RIPE's data shows the
allocation to C&W not the sub-allocation.

-----Burton